ci(publish): ship lazycodex as Node installer
